+++ src/main/java/com/grownited/repository/CityRepository.java
@@ -3,5 +3,5 @@ package com.grownited.repository;
 import com.grownited.entity.CityEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<CityEntity, Integer> {
+public interface CityRepository extends JpaRepository<CityEntity, Long> {
 }
