+++ src/main/java/com/grownited/repository/ClinicRepository.java
@@ -3,5 +3,5 @@ package com.grownited.repository;
 import com.grownited.entity.ClinicEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClinicRepository extends JpaRepository<ClinicEntity, Integer> {
+public interface ClinicRepository extends JpaRepository<ClinicEntity, Long> {
 }
